删除英文和中文的 README 文件，移除示例结果和相关图像；更新数据集处理代码以使用半精度浮点数，优化内存使用；调整模型和训练相关代码以支持半精度计算。

+++ datasets/cub_parts.py
@@ -348,7 +348,7 @@ class CUBDataset(Dataset):
         kp_norm, sfm_pose = self.normalize_kp(kp, sfm_pose, img_h, img_w)
 
         img = Image.fromarray(np.asarray(img, np.uint8))
-        mask = np.asarray(mask, np.float32)
+        mask = np.asarray(mask, np.float16)
         return img, kp_norm, mask, sfm_pose, img_path
 
     def normalize_kp(self, kp, sfm_pose, img_h, img_w):
